requirements.txt
streamlit>=1.22.0
pandas>=1.4.0
numpy>=1.21.0
openpyxl>=3.0.10
xlsxwriter>=3.0.3
plotly>=5.11.0
Pillow>=9.0.0


